Try and fix reported error in sidekiq
Have another go at fixing issues when not in a web request
context (e.g. sidekiq)

This has not been tested very well (no automated tests) etc
so am putting in a separate branch for the moment.

Just a quick hack to see if it helps pioneer-user who is
having issues